refactor: update directory name
