+++ src/Views/Components/ListTable/ProductRecommendations/style.module.scss
